create initial board state and draw pieces
Starting a local game will now draw a board, call the backend chess engine to generate the initial board state as a JSON object, and will draw the pieces in the correct places on the board. Pieces also each have their own click listener that for now will just report information about themselves to the console.