Restructured board states and implemented "generateBoardState()"
Restructured the way a boardState object is organized. This involved changes to nearly every function written so far, but is FAR more readable. Implemented generateBoardState(). Added some functionality to local games where you can pass a room code as a GET parameter and it will recall that specific game instead of generating an entirely new game every time the page is refreshed.