Fixed most of the stuff I broke last night
Turns out doing something like newDict = oldDict will cause changes in either dictionary to be reflected in the other. This was causing a lot of weird problems in the simulateMove() function I wrote last night. That's been corrected, but now when a player is put in check they are unable to move afterwards. Need to figure out what's causing that.