Implemented Promotions and Castling
Pawns that have made it across the board will now automatically be promoted to Queen. May need to add a UI in the future to allow the user to pick the piece to promote to. Castling is now possible for either King in either direction and will respect rules about not moving through check. Click handlers on the frontend will now propagate through the pieces to the cells behind them if and only if the user is already checking their moves. (No more having to click around the corners of a cell to capture a piece, when the piece was blocking the click)