Implemented En Passant
Implemented En Passant in checkPieceMoves() and generateBoardState(). Updated simulateMove() to check for En Passant, promotions, and castling. Fixed a bug in isInCheck() where being in check by a pawn wouldn't register as a check. Removed checks in isInCheck() for opposing Kings, since a King cannot place another King in check directly. Fixed a bug in checkPieceMoves() that would allow certain moves to be made if a players King was in check, even if that player would remain in check after the move. Added an argument to generateBoardState() to get only a certain subset of moves, which will be used later on for replays.