Finished Styling home page
Added a bunch of CSS to make the home page look a little nicer. Now displays active games, both local & network in separate columns. Displays a big banner across the top to show the number of wins, draws, and losses a user has. Visiting this page without being logged in will redirect to the login page.