Fixed a few bugs and wrote some documentation
fixed a bug where entering a lowercase room code in the "Join Game" form returns an error.

trying to view "reviewGame" with an invalid room code now redirects to the "review" page.

attempting to join a network game that's already full now redirects to the 'play' page with an error message for the user instead of crashing the site.